Move pkg/k8sutil to internal/k8sutil
Remove all public APIs from SDK to prepare for v1.0.0
